worker/requirements.txt
celery==5.4.0
redis==5.0.8
qdrant-client==1.11.0
sentence-transformers==3.1.1
psycopg2-binary==2.9.9
pypdf==4.3.1
httpx==0.27.2
unstructured==0.15.13
